Use protocol factory instead of resolver
